requirements.txt
flask==3.1.1
requests==2.32.3
pymongo==4.13.0
PyJWT==2.10.1
dotenv==0.9.9
flasgger==0.9.7.1